refactor: ミニマップ描画処理を分割
